backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
